rename: useGetChats.ts -> useHandleChats.ts #134
