+++ frontend/src/components/models/task/TaskCommentArea.tsx
@@ -5,7 +5,7 @@ import { UserAvatarIcon } from 'components/ui/avatar/UserAvatarIcon'
 import { SmallPopover } from 'components/ui/popup/SmallPopover'
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
 import { faEllipsisH } from '@fortawesome/free-solid-svg-icons'
-import { useHandleChats } from 'hooks/useGetChats'
+import { useHandleChats } from 'hooks/useHandleChats'
 import { usePopover } from 'hooks/usePopover'
 import { convertIntoRGBA } from 'utils/color/convertIntoRGBA'
 import { calculateMinSizeBasedOnFigma } from 'utils/calculateSizeBasedOnFigma'
